ci: disable build docs temporarily (#63)
the dependencies required to build the docs weren't version pinned,
and currently they're updated at upstream. causing the build to fail
for an unknown reason.
other branches in development require the ci to be working properly,
and thus disabling them for now until the root problem is figured out,
the dependencies are pinned, and the ci is fixed.